+++ src/lib.rs
@@ -147,7 +147,7 @@ pub use timecode::{
     parse_tmcd_sample_description, Tmcd, TMCD_FLAG_24_HOUR, TMCD_FLAG_COUNTER,
     TMCD_FLAG_DROP_FRAME, TMCD_FLAG_NEGATIVES_OK,
 };
-pub use track::{SampleDescription, Track, TrackRef, TrackRefKind};
+pub use track::{SampleDescription, SoundV1, Track, TrackRef, TrackRefKind};
 pub use track_group::{
     parse_track_group_type, parse_trgr, TrackGroupTypeEntry, TRACK_GROUP_TYPE_MSRC,
 };